test scene exit button from hello world scene alarm off event in timer ui in addition to timer update event hello world scene not needed anymore?

+++ Include/WkCocos/Timer/Events/TimerUpdate.h
@@ -16,9 +16,9 @@ namespace WkCocos
 				, time_left(time)
 				{}
 
-				double getTimeLeft() const
+				int getTimeLeft() const
 				{
-					return time_left;
+					return (int) time_left;
 				}
 
 				entityx::Entity timer;
